[Caching] Harden per-construct split gate and move it to its own file
Addresses the Codex review on #880.

Contact area (AGENTS.md): move the per-construct frontend split -- segmentation,
backward slice, recompute-safety gate, and orchestration -- out of the central
compile_to_offloads pass into a new transforms/split_frontend_per_construct.cpp
that exposes a single irpass::maybe_split_frontend_per_construct(). The central
pass now keeps only the one gated call.

Recompute-safety gate, two soundness fixes:

- Loop-carried local shared across constructs. The old check compared readers
  against the UNION of writer constructs, so two constructs that both
  read-modify-write the same local were each "covered" by the union and wrongly
  accepted -- the second still consumes the value the first produced. Now a
  loop-produced local is safe only when every access (read or loop-write) stays
  within a single construct.

- Field load snapshotted before a later construct. The backward slice recomputes
  a serial field/ndarray load into every consuming construct; if a construct
  between the load and its consumer writes global memory, the recomputed load
  observes the mutation instead of the source-order snapshot (e.g.
  base = x[0]; for: x[0] = 2; for i: y[i] = base). Now falls back when a
  recomputed load is shadowed by an intervening global write. Field identity is
  not tracked, so this is conservative.

Tests: add fallback + correctness regression tests for both miscompiles
(test_..._fallback_field_load_shadowed, test_..._fallback_carried_rmw_local).
