+++ tests/python/test_per_offload_cache.py
@@ -1,11 +1,12 @@
 """Per-construct FRONTEND split (no reuse tier).
 
 `compile_to_offloads` splits the frontend (simplify / merge_global_ptrs / offload) to run per top-level construct
-instead of once over the whole kernel, isolating each construct by its backward slice, and falls back to the
-whole-kernel path for kernels that are not recompute-safe (autodiff, mesh-for, or a local produced inside one
-construct and consumed by another). This PR ships the split WITHOUT a reuse tier, so it recompiles every construct on
-every compile; the reuse (a disk manifest keyed by `get_hashed_per_construct_cache_key`) is added by the cross-process
-cache PR.
+instead of once over the whole kernel, isolating each construct by its backward slice (in
+`transforms/split_frontend_per_construct.cpp`), and falls back to the whole-kernel path for kernels that are not
+recompute-safe: autodiff, mesh-for, a loop-produced local shared across constructs, or a serial field load that a
+later construct would recompute after an intervening construct mutated global memory. This PR ships the split WITHOUT a
+reuse tier, so it recompiles every construct on every compile; the reuse (a disk manifest keyed by
+`get_hashed_per_construct_cache_key`) is added by the cross-process cache PR.
 
 These tests assert the split's STRUCTURE and CORRECTNESS, not reuse:
  - the split fires for recompute-safe kernels and enumerates the expected constructs, with
@@ -106,3 +107,54 @@ def test_per_construct_frontend_split_fallback_not_recompute_safe() -> None:
 
     # Correctness of the whole-kernel fallback path.
     assert np.allclose(y.to_numpy(), float(np.arange(_N).sum()), atol=1e-2), y.to_numpy()
+
+
+@test_utils.test(arch=[qd.cpu, qd.cuda], offline_cache=False)
+def test_per_construct_frontend_split_fallback_field_load_shadowed() -> None:
+    # `base` snapshots x[0] BEFORE the first loop overwrites x[0]; a later construct consumes `base`. The backward
+    # slice would recompute the `x[0]` load into that later construct, where it would observe the mutated value instead
+    # of the source-order snapshot -- so the kernel is NOT recompute-safe and must fall back to the whole-kernel path.
+    @qd.kernel
+    def kernel_shadow(x: qd.types.ndarray(), y: qd.types.ndarray()) -> None:
+        base = x[0]
+        for i in range(_N):
+            x[i] = 2.0
+        for i in range(_N):
+            y[i] = base
+
+    x = qd.ndarray(qd.f32, shape=(_N,))
+    y = qd.ndarray(qd.f32, shape=(_N,))
+    x.from_numpy(np.full(_N, 7.0, dtype=np.float32))
+    kernel_shadow(x, y)
+
+    obs = kernel_shadow._primal.per_offload_cache_observations
+    assert obs.frontend_constructs_total == -1, obs
+
+    # `base` must be the ORIGINAL x[0] (7.0), not the value the first loop wrote (2.0). A split that recomputed the
+    # load after the store would wrongly yield 2.0.
+    assert np.allclose(y.to_numpy(), 7.0, atol=1e-2), y.to_numpy()
+
+
+@test_utils.test(arch=[qd.cpu, qd.cuda], offline_cache=False)
+def test_per_construct_frontend_split_fallback_carried_rmw_local() -> None:
+    # Two constructs each read-modify-write the same local `s`, and the second also stores it. The second construct
+    # depends on the value the first produced, so it is not recomputable per-construct (its slice would drop the first
+    # loop and restart `s` from the serial init). Checking readers against the *union* of writer constructs would wrongly
+    # accept this (both readers are also writers), so the gate must reject it and fall back.
+    @qd.kernel
+    def kernel_carry(out: qd.types.ndarray()) -> None:
+        s = 1
+        for i in range(1):
+            s += 1
+        for i in range(1):
+            s += 1
+            out[0] = s
+
+    out = qd.ndarray(qd.i32, shape=(1,))
+    kernel_carry(out)
+
+    obs = kernel_carry._primal.per_offload_cache_observations
+    assert obs.frontend_constructs_total == -1, obs
+
+    # s: 1 -> 2 (loop A) -> 3 (loop B). A split that dropped loop A from B's slice would wrongly produce 2.
+    assert int(out.to_numpy()[0]) == 3, out.to_numpy()
